Converted the new APIs to internal ones in linux.net.
