--- kura/org.eclipse.kura.api/src/main/java/org/eclipse/kura/net/modem/GatewayModemDriver.java
@@ -1,15 +1,0 @@
-package org.eclipse.kura.net.modem;
-
-import org.eclipse.kura.KuraException;
-
-/**
- * @since 2.0
- */
-public interface GatewayModemDriver {
-    
-    public void turnModemOff(String vendor, String product) throws KuraException;
-    
-    public void turnModemOn(String vendor, String product) throws KuraException;
-
-    public void resetModem(String vendor, String product) throws KuraException;
-}
