+++ kura/org.eclipse.kura.linux.debian.provider/META-INF/MANIFEST.MF
@@ -13,6 +13,7 @@ Import-Package: org.apache.commons.io;version="2.4.0",
  org.eclipse.kura.core.linux.util;version="[1.1,2.0)",
  org.eclipse.kura.core.net;version="1.0.5",
  org.eclipse.kura.core.net.util;version="1.0.0",
+ org.eclipse.kura.internal.linux.net;version="1.0.0",
  org.eclipse.kura.linux.net.dhcp;version="1.0.0",
  org.eclipse.kura.linux.net.dns,
  org.eclipse.kura.net;version="[2.0,3.0)",
