+++ kura/org.eclipse.kura.linux.debian.provider/OSGI-INF/netinterfaceserializer.xml
@@ -1,0 +1,21 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+
+    Copyright (c) 2018 Eurotech and/or its affiliates
+
+     All rights reserved. This program and the accompanying materials
+     are made available under the terms of the Eclipse Public License v1.0
+     which accompanies this distribution, and is available at
+     http://www.eclipse.org/legal/epl-v10.html
+
+    Contributors:
+      Eurotech
+
+-->
+<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" enabled="true" immediate="true" name="org.eclipse.kura.internal.linux.net.config.NetInterfaceConfigSerializationService">
+   <implementation class="org.eclipse.kura.internal.linux.net.config.NetInterfaceConfigSerializationServiceImpl"/>
+   <service>
+      <provide interface="org.eclipse.kura.internal.linux.net.NetInterfaceConfigSerializationService"/>
+   </service>
+   <property name="service.pid" value="org.eclipse.kura.internal.linux.net.config.NetInterfaceConfigSerializationService"/>
+</scr:component>
