+++ kura/org.eclipse.kura.linux.net/META-INF/MANIFEST.MF
@@ -34,14 +34,19 @@ Import-Package: javax.microedition.io,
  org.osgi.service.io;version="1.0.0",
  org.osgi.util.tracker;version="[1.5.0,2.0.0)",
  org.slf4j;version="1.6.4"
-Export-Package: org.eclipse.kura.linux.net;version="1.0.9";
+Export-Package: org.eclipse.kura.internal.board;version="1.0.0";x-internal:=true,
+ org.eclipse.kura.internal.linux.net;version="1.0.0";x-internal:=true,
+ org.eclipse.kura.internal.linux.net.dns;version="1.0.0";x-internal:=true,
+ org.eclipse.kura.internal.linux.net.modem;version="1.0.0";x-internal:=true,
+ org.eclipse.kura.internal.linux.net.wifi;version="1.0.0";x-internal:=true,
+ org.eclipse.kura.linux.net;version="1.0.9";
   uses:="org.osgi.service.component,
    org.osgi.service.event,
    org.eclipse.kura.net,
    org.eclipse.kura.net.modem,
    org.eclipse.kura.usb",
  org.eclipse.kura.linux.net.dhcp;version="1.0.0";uses:="org.eclipse.kura.net.dhcp",
- org.eclipse.kura.linux.net.dns,
+ org.eclipse.kura.linux.net.dns;version="1.0.1";uses:="org.eclipse.kura.net,org.eclipse.kura.net.dns",
  org.eclipse.kura.linux.net.iptables;version="1.0.1";uses:="org.eclipse.kura.net",
  org.eclipse.kura.linux.net.modem;version="1.1.1";uses:="org.osgi.service.event,org.eclipse.kura.usb",
  org.eclipse.kura.linux.net.ppp;version="1.0.0",
