+++ kura/org.eclipse.kura.linux.sysv.provider/META-INF/MANIFEST.MF
@@ -10,6 +10,7 @@ Bundle-ClassPath: .
 Bundle-ActivationPolicy: lazy
 Import-Package: org.eclipse.kura;version="[1.1,2.0)",
  org.eclipse.kura.core.linux.util;version="[1.1,2.0)",
+ org.eclipse.kura.internal.linux.net.dns;version="[1.0,2.0)",
  org.eclipse.kura.net;version="[2.0,3.0)",
  org.eclipse.kura.net.dns;version="[1.0,2.0)",
  org.osgi.service.component;version="1.2.0",
