+++ kura/org.eclipse.kura.net.admin/META-INF/MANIFEST.MF
@@ -17,9 +17,13 @@ Import-Package: javax.microedition.io,
  org.eclipse.kura.core.net.modem;version="[1.0,2.0)",
  org.eclipse.kura.core.net.util;version="[1.0,2.0)",
  org.eclipse.kura.core.util;version="[1.0,2.0)",
+ org.eclipse.kura.internal.board;version="[1.0,2.0)",
+ org.eclipse.kura.internal.linux.net;version="[1.0,2.0)",
+ org.eclipse.kura.internal.linux.net.dns;version="[1.0,2.0)",
+ org.eclipse.kura.internal.linux.net.wifi;version="[1.0,2.0)",
  org.eclipse.kura.linux.net;version="[1.0,2.0)",
  org.eclipse.kura.linux.net.dhcp;version="[1.0,2.0)",
- org.eclipse.kura.linux.net.dns,
+ org.eclipse.kura.linux.net.dns;version="[1.0,2.0)",
  org.eclipse.kura.linux.net.iptables;version="[1.0,2.0)",
  org.eclipse.kura.linux.net.modem;version="[1.1,2.0)",
  org.eclipse.kura.linux.net.ppp;version="[1.0,2.0)",
