+++ kura/org.eclipse.kura.net.admin/OSGI-INF/dnsMonitor.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!--
 
-    Copyright (c) 2011, 2016 Eurotech and/or its affiliates
+    Copyright (c) 2011, 2018 Eurotech and/or its affiliates
 
      All rights reserved. This program and the accompanying materials
      are made available under the terms of the Eclipse Public License v1.0
@@ -18,5 +18,5 @@
       <provide interface="org.eclipse.kura.net.dns.DnsMonitorService"/>
    </service>
    <reference bind="setNetworkConfigurationService" cardinality="1..1" interface="org.eclipse.kura.net.admin.NetworkConfigurationService" name="NetworkConfigurationService" policy="static" unbind="unsetNetworkConfigurationService"/>
-   <reference bind="setDnsServerService" cardinality="1..1" interface="org.eclipse.kura.net.dns.DnsServer" name="DNSService" policy="static" unbind="unsetDnsServerService"/>
+   <reference bind="setDnsServerService" cardinality="1..1" interface="org.eclipse.kura.internal.linux.net.dns.DnsServerService" name="DNSService" policy="static" unbind="unsetDnsServerService"/>
 </scr:component>
