+++ kura/org.eclipse.kura.net.admin/OSGI-INF/networkAdmin.xml
@@ -23,6 +23,6 @@
    <reference bind="setConfigurationService" cardinality="1..1" interface="org.eclipse.kura.configuration.ConfigurationService" name="ConfigurationService" policy="static" unbind="unsetConfigurationService"/>
    <reference bind="setNetworkConfigurationService" cardinality="1..1" interface="org.eclipse.kura.net.admin.NetworkConfigurationService" name="NetworkConfigurationService" policy="static" unbind="unsetNetworkConfigurationService"/>
    <reference bind="setFirewallConfigurationService" cardinality="1..1" interface="org.eclipse.kura.net.admin.FirewallConfigurationService" name="FirewallConfigurationService" policy="static" unbind="unsetFirewallConfigurationService"/>
-   <reference bind="setDnsServerService" cardinality="1..1" interface="org.eclipse.kura.net.dns.DnsServer" name="DNSService" policy="static" unbind="unsetDnsServerService"/>
-   <reference bind="setWifiUtilsService" cardinality="0..1" interface="org.eclipse.kura.net.wifi.WifiUtils" name="WifiUtils" policy="static" unbind="unsetWifiUtilsService"/>
+   <reference bind="setDnsServerService" cardinality="1..1" interface="org.eclipse.kura.internal.linux.net.dns.DnsServerService" name="DNSService" policy="static" unbind="unsetDnsServerService"/>
+   <reference bind="setWifiDriverService" cardinality="0..1" interface="org.eclipse.kura.internal.linux.net.wifi.WifiDriverService" name="WifiDriverService" policy="static" unbind="unsetWifiDriverService"/>
 </scr:component>
