+++ kura/org.eclipse.kura.net.admin/OSGI-INF/wifiMonitor.xml
@@ -26,5 +26,5 @@
    <reference bind="setNetworkService" cardinality="1..1" interface="org.eclipse.kura.net.NetworkService" name="NetworkService" policy="static" unbind="unsetNetworkService"/>
    <reference bind="setNetworkAdminService" cardinality="1..1" interface="org.eclipse.kura.net.NetworkAdminService" name="NetworkAdminService" policy="static" unbind="unsetNetworkAdminService"/>
    <reference bind="setNetworkConfigurationService" cardinality="1..1" interface="org.eclipse.kura.net.admin.NetworkConfigurationService" name="NetworkConfigurationService" policy="static" unbind="unsetNetworkConfigurationService"/>
-   <reference bind="setWifiUtilsService" cardinality="0..1" interface="org.eclipse.kura.net.wifi.WifiUtils" name="WifiUtils" policy="static" unbind="unsetWifiUtilsService"/>
+   <reference bind="setWifiDriverService" cardinality="0..1" interface="org.eclipse.kura.internal.linux.net.wifi.WifiDriverService" name="WifiDriverService" policy="static" unbind="unsetWifiDriverService"/>
 </scr:component>
