+++ kura/org.eclipse.kura.net.admin/src/main/java/org/eclipse/kura/net/admin/monitor/WifiMonitorServiceImpl.java
@@ -11,6 +11,7 @@
  *******************************************************************************/
 package org.eclipse.kura.net.admin.monitor;
 
+import java.io.IOException;
 import java.net.InetAddress;
 import java.util.ArrayList;
 import java.util.Collection;
@@ -33,6 +34,8 @@ import org.eclipse.kura.KuraException;
 import org.eclipse.kura.core.net.AbstractNetInterface;
 import org.eclipse.kura.core.net.NetworkConfiguration;
 import org.eclipse.kura.core.net.WifiInterfaceConfigImpl;
+import org.eclipse.kura.internal.board.BoardPowerState;
+import org.eclipse.kura.internal.linux.net.wifi.WifiDriverService;
 import org.eclipse.kura.linux.net.route.RouteService;
 import org.eclipse.kura.linux.net.route.RouteServiceImpl;
 import org.eclipse.kura.linux.net.util.IScanTool;
@@ -64,7 +67,6 @@ import org.eclipse.kura.net.wifi.WifiClientMonitorService;
 import org.eclipse.kura.net.wifi.WifiConfig;
 import org.eclipse.kura.net.wifi.WifiInterfaceAddressConfig;
 import org.eclipse.kura.net.wifi.WifiMode;
-import org.eclipse.kura.net.wifi.WifiUtils;
 import org.osgi.service.component.ComponentContext;
 import org.osgi.service.event.Event;
 import org.osgi.service.event.EventAdmin;
@@ -99,7 +101,7 @@ public class WifiMonitorServiceImpl implements WifiClientMonitorService, EventHa
     private NetworkConfiguration currentNetworkConfiguration;
     private NetworkConfiguration newNetConfiguration;
 
-    private WifiUtils wifiUtils;
+    private WifiDriverService wifiDriverService;
 
     public void setNetworkService(NetworkService networkService) {
         this.networkService = networkService;
@@ -139,12 +141,12 @@ public class WifiMonitorServiceImpl implements WifiClientMonitorService, EventHa
         this.netConfigService = null;
     }
 
-    public void setWifiUtilsService(WifiUtils wifiUtils) {
-        this.wifiUtils = wifiUtils;
+    public void setWifiDriverService(WifiDriverService wifiDriverService) {
+        this.wifiDriverService = wifiDriverService;
     }
 
-    public void unsetWifiUtilsService(WifiUtils wifiUtils) {
-        this.wifiUtils = null;
+    public void unsetWifiDriverService(WifiDriverService wifiUtils) {
+        this.wifiDriverService = null;
     }
 
     protected void activate(ComponentContext componentContext) {
@@ -217,8 +219,14 @@ public class WifiMonitorServiceImpl implements WifiClientMonitorService, EventHa
     }
 
     protected boolean isWifiDeviceOn(String interfaceName) {
-        if (this.wifiUtils != null) {
-            return this.wifiUtils.isWifiDeviceOn(interfaceName);
+        if (this.wifiDriverService != null) {
+            BoardPowerState wifiState;
+            try {
+                wifiState = this.wifiDriverService.getState(interfaceName);
+            } catch (IOException e) {
+                wifiState = BoardPowerState.OFF;
+            }
+            return wifiState == BoardPowerState.ON;
         }
         return false;
     }
@@ -281,8 +289,8 @@ public class WifiMonitorServiceImpl implements WifiClientMonitorService, EventHa
                     // There are interfaces for which we need to initially check if
                     // the right kernel module is loaded for the desired mode.
                     // If not we treat the interface as if needing to be reconfigured.
-                    if (this.first
-                            && wifiUtils != null && !wifiUtils.isKernelModuleLoadedForMode(interfaceName, wifiConfig.getMode())) {
+                    if (this.first && this.wifiDriverService != null && !this.wifiDriverService
+                            .isKernelModuleLoadedForMode(interfaceName, wifiConfig.getMode())) {
                         logger.info("monitor() :: {} kernel module not suitable for WiFi mode {}", interfaceName,
                                 wifiConfig.getMode());
                         this.first = false;
@@ -665,13 +673,13 @@ public class WifiMonitorServiceImpl implements WifiClientMonitorService, EventHa
 
     private void reloadKernelModule(String interfaceName, WifiMode wifiMode) throws KuraException {
         logger.info("monitor() :: reload {} using kernel module for WiFi mode {}", interfaceName, wifiMode);
-        if (wifiUtils == null) {
+        if (this.wifiDriverService == null) {
             return;
         }
-        if (wifiUtils.isKernelModuleLoaded(interfaceName)) {
-            wifiUtils.unloadKernelModule(interfaceName);
+        if (this.wifiDriverService.isKernelModuleLoaded(interfaceName)) {
+            this.wifiDriverService.unloadKernelModule(interfaceName);
         }
-        wifiUtils.loadKernelModule(interfaceName, wifiMode);
+        this.wifiDriverService.loadKernelModule(interfaceName, wifiMode);
     }
 
     private void initializeMonitoredInterfaces(NetworkConfiguration networkConfiguration) throws KuraException {
@@ -917,7 +925,8 @@ public class WifiMonitorServiceImpl implements WifiClientMonitorService, EventHa
     @Override
     public int getSignalLevel(String interfaceName, String ssid) throws KuraException {
         int rssi;
-        if (wifiUtils != null && !wifiUtils.isKernelModuleLoadedForMode(interfaceName, WifiMode.INFRA)) {
+        if (this.wifiDriverService != null
+                && !this.wifiDriverService.isKernelModuleLoadedForMode(interfaceName, WifiMode.INFRA)) {
             logger.info("getSignalLevel() :: reload {} kernel module for WiFi mode {}", interfaceName, WifiMode.INFRA);
             reloadKernelModule(interfaceName, WifiMode.INFRA);
         }
@@ -1004,13 +1013,13 @@ public class WifiMonitorServiceImpl implements WifiClientMonitorService, EventHa
         return statuses;
     }
 
-    private boolean resetWifiDevice(String interfaceName) throws KuraException {
+    private boolean resetWifiDevice(String interfaceName) throws IOException {
         boolean ret = false;
-        if (isWifiDeviceOn(interfaceName) && wifiUtils != null) {
-            wifiUtils.turnWifiDeviceOff(interfaceName);
+        if (isWifiDeviceOn(interfaceName) && this.wifiDriverService != null) {
+            this.wifiDriverService.disable(interfaceName);
         }
-        if (isWifiDeviceReady(interfaceName, false, 10) && wifiUtils != null) {
-            wifiUtils.turnWifiDeviceOn(interfaceName);
+        if (isWifiDeviceReady(interfaceName, false, 10) && this.wifiDriverService != null) {
+            this.wifiDriverService.enable(interfaceName);
             ret = isWifiDeviceReady(interfaceName, true, 20);
         }
         return ret;
